[pipelines] sionna 1.0 ray tracing working
sionna 1.0 needs to mature a bit first. it has several things broken and this would take more
effort than I want to invest now.
Also, unifying the actions for sionna <1 and >=1 is not trivial.
This will need some work on the exporter (started already) and on the converter (not started).
But Sionna 1.0 consumes way less memory and it's noticeably faster!
Worth doing, just not now.

Also, it has less dependencies, but it still only runs on Linux (on GPU).
LLVM can be installed to run on Windows too, but slow. Worth for prototyping though.

Additionally, sionna needs more coverage:
	- materials, building labels, roads

And the whole pipeline needs the CSV loop step before